Change: Recoder2KeYTypeConverter fail-fast when no sort for primitive type is found
If RecodeR tries to translate a primitive type and
cannot associate a known sort, it throws an
exception rather than creating silently a new sort.

* To allow reading programs that may contain floats
or doubles, it was necessary to implement stubs
of their LDTs.

* To avoid slowing down tests the old behaviour is
still in place when the conversion has been initiated
(implicitly) by junit.textui.TestRunner