Follow-up workaround for bug #950:
  The fix chosen for #950 required access to a package private method and extends therefore the
recoder.service package. The problem is that eclipse uses different classloaders for different
bundles/plugins and throws an illegal access error even though the classes are logically in the
same package.

This workaround should be replaced in the near future by a new release of our recoderKey.jar
making the method in question (DefaultProgramModelInfo#registerSubtype) protected.