Fix for determining the corrrect attribute name
Problem:
   user defined attributes are named className::$attributeName
   implicit ones are named className::<attributeName>
Note the missing $ for the implicit attributes.

This version (joined work with Wojciech) introduces a special case to take care of this.

The nicer solution would be to rename implicit attribute sto  className::$<attributeName>
as this would not require special treatment. But such a changes is not Friday stuff.