Do not only test for command line option startsWith("AUTO")
The only valid command line options for the user interface mode are
"AUTO" and "AUTO_LOADONLY" and not "AUTO_INTERACTIVE" for instance. Thus
test for equality instead of common prefix with one of the valid command
line options in Main#evaluateOptions.

The equality must be checked the first time (when setting uiMode)
because when the user interface is instantiated the implementation
expects uiMode to be INTERACTIVE when it is not "AUTO" or
"AUTO_LOADONLY".
The second time (when instantiating the user interface) it would be
correct to still check the prefix only but with the equality check how
uiMode was set is independent from the instantiation code and it can be
extracted from evaluateOptions.