Add convenience factories declare to KeYJavaASTFactory once more
The new methods declare(modifier, ...) and declare(modifiers, ...)
overload declare(ProgramVariable, Expression, KeYJavaType) to be able to
define modifiers for LocalVariableDeclarations.

From now on the original declare(ProgramVariable, Expression,
KeYJavaType) is implemented using declare(modifiers, ...) with an empty
array and declare(ProgramVariable, KeYJavaType) is implemented using the
former with a null init argument.