Add convenience factories methodCall to KeYJavaASTFactory
The two implementations of methodCall create a new MethodReference
object from a method name and a variable. One of them additionally takes
a list of arguments to be passed to the method call and the other one
wraps around the first one passing its arguments and an empty arguments
list to it.