Overload fieldReference to already accept ProgramVariable
KeYJavaASTFactory::fieldReference now also accepts ProgramVariables
instead of String names of fields. Therefore, Services need not be
specified anymore. Moreover, the new signatures expects not only an
Expression but a ReferencePrefix already.

Implement the original fieldReference using the new fieldReference
signature after the ProgramVariable is found and the Expression is
parenthesized.