Overload methodCall to accept KeYJavaType as prefix
KeYJavaASTFactory::methodCall now also accepts specifically KeYJavaType
as prefix for the to be created method call. A TypeReference, which
implements ReferencePrefix, is created from the type argument and the
ReferencePrefix aware methodCall signatures are used to actually create
the MethodCall instances.

Two versions of the new signature are implemented, one takes a list of
arguments for the to be created method call from the caller and the
other implicitly creates an empty argument list.