Rule application restrictions "\succedentPolarity" and "\antecedentPolarity" added.
If a rewrite taclet is marked with "\succedentPolarity", it will only be applied under the condition that, if the surrounding formula has been decomposed completely, the find-term will NOT appear on the ANTEcedent. The formula "wellformed(h)" in  "==> wellformed(h)" or in "wellformed(h) -> (inv(h) = inv(h2)) ==>" or in "\if(b) \then(!wellformed(h)) \else(!wellformed(h2)) ==>" has succedent polarity. The formula "wellformed(h)" in "wellformed(h) <-> wellformed(h2) ==>" has NO succedent polarity.

"\antecedentPolarity" is defined analog.