[Test] Increase the number of steps for TestProofStarter
This test failed (intermittently) because of a proof cut occurring
sometimes to late. At the moment the proof requires 50000-51000 steps,
while with the cut occurring at the right place these are < 15000.

The previous checkins did not change anything in the strategies, so the
larger proof is most likely due to some order of the queue and identical
costs for several rule applications.

The fix is only a workaround. the right thing would be to improve the
strategies systematically. When lowering the standard costs for cuts to
75 instead of 100. The shorter proof is found. But that is magic with
numbers and for an actual fix a more systematic treatment or setting of
costs is required. That is why this patch just increases the max number
of steps.