[Fix] Fixed an issue with timeout and time reporting
The timeout option should not have worked since long ago, when I changed
the counters to nano. The reasoning was that (if available) high
performance counters would be used. But that screwed up other issues like
timeout for strategies and lead sometimes to overflow error. This commit
reverses that mistake. Mea culpa!