Loading of single Java files.
This commit offers a new action for the user: the loading of single Java
files.

If you load/open a Java in the traditional way, KeY considers the path
of the Java as part of the classpath. Hence, all Java files found
recursively under the given folder are parsed.

This commit adds a new action, and extends the core facilities to avoid
this behaviour.

# Why is this needed?

- The default case in KeY is loading a Java file, that contains all
  resources and only depends on the JDK.

- The traditional behaviour is misleading for KeY newbies.