avoid creating proof obligations in ProofManagementDialog It is unnecessary to create proof obligations in most cases there and creating some proof obligations changes state shared with a large part of KeY (see #1715).
This does *not* fix #1715, just makes it unlikely to encounter as
you now first have to start an information flow proof and then in a
proof for another class try to cut on `self`, in the cut dialog KeY will
think `self` comes from the information flow proof.

As a workaround for #1715 do all information flow proofs you want to do
last.