Fast fix: Do not run pipeline defined by the Jenkinsfile on "Master" worker
Jenkinsfile is used for the pipeline definition for merge requests. Currently, this pipeline is also executed on the "Master" Jenkins worker. Due to wrong resource handling this threaten the availability of the Jenkins control UI (which runs on the same machine with same permissions and user).

This MR disables this special worker node for Jenkins build on merge requests.