Add bit extraction function `bitAt` for JavaDL ints
The `bitAt` returns the value of a single bit (0 or 1) from an unbounded
JavaDL int, assuming two's complement representation. For negative
indices, the result is zero.

Introduced for definitions of `binaryAnd`, `binaryOr`, and `binaryXOr`
in a later commit. Developed as part of my bachelor's thesis "Handling
Bitwise Operations in Deductive Program Verification: Case Study
java.util.BitSet with KeY" (2023).