Make TermLabelManager independent of specific TermLabel instances
The TermLabelManager provides a general framework for managing term labels and should not have dependencies to specific instances (otherwise we would need to touch it anytime we add a new TermLabel. All TermLabel specific code should reside in the classes implementing that termlabel or its refactorings)

I deleted the code as with the proposal implemented in this PR #3316 it would be no longer needed. That is why I decided against the other options: (i) moving the code into one of the OriginTermLabel* classes or generalising it, such that it could be used by any termlabel.